Use hooks to manage assessment state:
