+++ src/components/screen-components/Assessment/Slides/AssessmentSlide1.tsx
