+++ src/components/screen-components/Assessment/Slides/AssessmentSlide3.tsx
