+++ src/components/screen-components/Assessment/Slides/AssessmentSlide4.tsx
